+++ src/scripts/main/dialog/index.js
@@ -1,4 +1,4 @@
-import { handleScrollbar, openComponent, closeComponent } from "../utils";
+import { openComponent, closeComponent } from "../utils";
 
 const SELECTORS = {
 	DIALOG: ".dialog",
@@ -96,9 +96,8 @@ class Dialog {
 
 	open() {
 		this.focusedElemBeforeDialog = document.activeElement;
-		handleScrollbar();
+
 		this.isOpen = true;
-		this._el.style.display = "flex";
 		document.body.classList.add("dialog-open");
 		this.lastTabStop.focus();
 
@@ -106,10 +105,7 @@ class Dialog {
 	}
 
 	close() {
-		handleScrollbar();
 		this.isOpen = false;
-		this._el.style.display = "none";
-		this._el.classList.remove("d-flex");
 		document.body.classList.remove("dialog-open");
 		this.focusedElemBeforeDialog ? this.focusedElemBeforeDialog.focus() : null;
 
@@ -125,6 +121,31 @@ const _createDialog = (dialogQuery) => {
 	return dialogObject;
 };
 
+// MARK: script for <dialog> element
+
+const _activateDialogElement = (dialog) => {
+	const dialogInvoker = document.querySelector(
+		`button[data-dialog-open="${dialog.id}"]`,
+	);
+	const closeDialogButtons = dialog.querySelectorAll(
+		"button[data-dialog-close]",
+	);
+
+	// add event listener on dialogInvoker, it should call dialog.showModal()
+	dialogInvoker.addEventListener("click", () => {
+		dialog.showModal();
+	});
+
+	// add event listener on dialogs close button, it should call dialog.close()
+	[...closeDialogButtons].map((closeBtn) =>
+		closeBtn.addEventListener("click", () => {
+			dialog.close();
+		}),
+	);
+
+	return dialog;
+};
+
 const init = (id) => {
 	if (id) {
 		const dialog = document.getElementById(id);
@@ -135,23 +156,37 @@ const init = (id) => {
 			return null;
 		}
 
-		return _createDialog(dialog);
+		if (dialog?.tagName.toLowerCase() === "dialog") {
+			return _activateDialogElement(dialog);
+		} else {
+			return _createDialog(dialog);
+		}
 	} else {
-		const dialogs = document.querySelectorAll(SELECTORS.DIALOG);
+		const dialogsLegacy = document.querySelectorAll(SELECTORS.DIALOG);
+		const modernDialogs = document.querySelectorAll("dialog");
 
-		if (!dialogs.length) {
+		if (![...dialogsLegacy, ...modernDialogs].length) {
 			console.warn("No dialogs found");
 
 			return null;
 		}
 
-		return [...dialogs].map((dialog) => _createDialog(dialog));
+		return [
+			...[...dialogsLegacy].map((dialog) => _createDialog(dialog)),
+			...[...modernDialogs].map((dialog) => _activateDialogElement(dialog)),
+		];
 	}
 };
 
-const open = (id) => openComponent(id, "dialog", _dialogs);
+const open = (id) =>
+	document.getElementById(id)?.tagName.toLowerCase() === "dialog"
+		? document.getElementById(id).showModal()
+		: openComponent(id, "dialog", _dialogs);
 
-const close = (id) => closeComponent(id, "dialog", _dialogs);
+const close = (id) =>
+	document.getElementById(id)?.tagName.toLowerCase() === "dialog"
+		? document.getElementById(id).close()
+		: closeComponent(id, "dialog", _dialogs);
 
 export default {
 	init,