+++ RELEASE_NOTES.md
@@ -17,8 +17,23 @@
     - It is now possible to add a subtext inside the checkmark variant (use span.subtext)
 
 - Buttons
+
   - minor non-breaking UI updates (height for size large, unify hover state, modify active & focus-visible UI, and more)
   - add Danish MitId button style (you can use either `i.bank-id-dk` OR `i.mitid-dk`, both are supported)
+
+- Cards
+
+  - fix style .cards-wide on mobile
+
+- Dialog
+
+  - UI update
+  - new syntax, leveraging the standard <dialog> tag
+  - silent support for the old syntax until the next major release
+  - add transitions
+
+- Buttons
+  - secondary buttons get a transparent background-color
 
 ## Technical changes